+++ src/tools/debug/DumpWeb.php
@@ -158,7 +158,7 @@ abstract class DumpWeb extends DumpMaster {
         echo $propertyType, ' => ', $propertyValue ? 'true' : 'false', $property->getDocComment();
         break;
       case 'integer' :
-      case 'float' :
+      case 'double' :
         echo $propertyType, ' => ', $propertyValue,  $property->getDocComment();
         break;
       case DumpMaster::OTRA_DUMP_TYPE_STRING :
@@ -247,7 +247,7 @@ abstract class DumpWeb extends DumpMaster {
         echo $paramType, $param ? ' true' : ' false', self::OTRA_DUMP_END_TEXT_BLOCK;
         break;
       case 'integer' :
-      case 'float' :
+      case 'double' :
         echo $param, '</span>', self::OTRA_DUMP_END_TEXT_BLOCK, '<br>';
         break;
       case 'NULL' : echo '<b>null</b><br>'; break;
